@@ -4,7 +4,7 @@ import 'package:one_context/src/components/one_basic_widget.dart';
 import 'package:one_context/src/components/one_notifier.dart';
 
 typedef Widget NotificationBuilder<T>(BuildContext context, T data);
-typedef void OnVisited<T>(BuildContext context, T data);
+typedef void OnVisited<T>(BuildContext context, T? data);
 
 /// Rebuild the widget below OneNotification
 /// It can be usually used to rebuild widget tree
@@ -26,43 +26,42 @@ typedef void OnVisited<T>(BuildContext context, T data);
 /// So, ever use a generic type to avoid unexpected behaviors.
 
 class OneNotification<T> extends StatefulWidget {
-  final NotificationBuilder<T> builder;
+  final NotificationBuilder<T?> builder;
   final bool stopBubbling;
   final bool rebuildOnNull;
   final bool rebuildOnData;
-  final OnVisited onVisited;
-  final T initialData;
-  final List<Type> types;
+  final OnVisited? onVisited;
+  final T? initialData;
+  final List<Type>? types;
 
   OneNotification({
-    @required this.builder,
+    required this.builder,
     this.initialData,
     this.onVisited,
     this.stopBubbling = false,
     this.rebuildOnNull = false,
     this.rebuildOnData = true,
     this.types,
-    Key key,
-  })  : assert(builder != null, 'builder is required!'),
-        super(key: key);
+    Key? key,
+  }) : super(key: key);
 
   /// Use it to reload a OneNotification widget children
   /// It use the type of NotificationPayload.data to decide if it has to make changes
-  static notify<T>(BuildContext context, {NotificationPayload<T> payload}) {
+  static notify<T>(BuildContext context, {NotificationPayload<T>? payload}) {
     OneNotifier(payload: payload)..dispatch(context);
   }
 
   /// Use it to reload the entire widget three, loose previous data of most top OneNotification found
   /// Cunrrently it recreate most top OneNotification
   static hardReloadRoot(BuildContext context) {
-    _OneNotificationState state =
+    _OneNotificationState? state =
         context.findRootAncestorStateOfType<_OneNotificationState>();
     state?._hardReload();
   }
 
   /// Cunrrently it rebuild most top OneNotification
   static softReloadRoot(BuildContext context) {
-    _OneNotificationState state =
+    _OneNotificationState? state =
         context.findRootAncestorStateOfType<_OneNotificationState>();
     state?._softReload();
   }
@@ -72,7 +71,7 @@ class OneNotification<T> extends StatefulWidget {
 }
 
 class _OneNotificationState<T> extends State<OneNotification<T>> {
-  T _data;
+  T? _data;
   Key _key = UniqueKey();
 
   @override
@@ -83,21 +82,21 @@ class _OneNotificationState<T> extends State<OneNotification<T>> {
           child: widget.builder(context, _data ?? widget.initialData),
         ),
         onNotification: (notification) {
-          widget.onVisited?.call(context, notification?.payload?.data);
+          widget.onVisited?.call(context, notification.payload?.data);
 
           /// Null data received
-          if (widget.rebuildOnNull && notification?.payload?.data == null) {
+          if (widget.rebuildOnNull && notification.payload?.data == null) {
             _softReload();
             return widget.stopBubbling;
 
             /// Data received and try match type
-          } else if (notification?.payload?.data?.runtimeType?.toString() ==
+          } else if (notification.payload?.data?.runtimeType.toString() ==
                   T.toString() ||
               dynamic is T) {
             _data = notification.payload?.data;
 
             if (widget.rebuildOnData) if (notification
-                    ?.payload?.forceHardHeload ==
+                    .payload?.forceHardHeload ==
                 true)
               _hardReload();
             else
