@@ -9,40 +9,39 @@ mixin DialogController {
   DialogController get dialog => this;
 
   /// The current context
-  BuildContext get context => OneContext().context;
+  BuildContext? get context => OneContext().context;
 
-  Future<T> Function<T>(
+  Future<T?> Function<T>(
       {bool barrierDismissible,
-      Widget Function(BuildContext) builder,
-      bool useRootNavigator}) _showDialog;
-  Future<T> Function<T>(
-      {Widget Function(BuildContext) builder,
-      Color backgroundColor,
-      double elevation,
-      ShapeBorder shape,
-      Clip clipBehavior,
+      required Widget Function(BuildContext) builder,
+      bool useRootNavigator})? _showDialog;
+  Future<T?> Function<T>(
+      {required Widget Function(BuildContext) builder,
+      Color? backgroundColor,
+      double? elevation,
+      ShapeBorder? shape,
+      Clip? clipBehavior,
       bool isScrollControlled,
       bool useRootNavigator,
-      bool isDismissible}) _showModalBottomSheet;
+      bool isDismissible})? _showModalBottomSheet;
   ScaffoldFeatureController<SnackBar, SnackBarClosedReason> Function(
-      SnackBar Function(BuildContext) builder) _showSnackBar;
+      SnackBar Function(BuildContext?) builder)? _showSnackBar;
   PersistentBottomSheetController<T> Function<T>(
       {Widget Function(BuildContext) builder,
-      Color backgroundColor,
-      double elevation,
-      ShapeBorder shape,
-      Clip clipBehavior}) _showBottomSheet;
+      Color? backgroundColor,
+      double? elevation,
+      ShapeBorder? shape,
+      Clip? clipBehavior})? _showBottomSheet;
 
   /// Displays a Material dialog above the current contents of the app, with
   /// Material entrance and exit animations, modal barrier color, and modal
   /// barrier behavior (dialog is dismissible with a tap on the barrier).
-  Future<T> showDialog<T>(
-      {@required Widget Function(BuildContext) builder,
+  Future<T?> showDialog<T>(
+      {required Widget Function(BuildContext) builder,
       bool barrierDismissible = true,
       bool useRootNavigator = true}) async {
-    assert(builder != null);
     if (!(await _contextLoaded())) return null;
-    return _showDialog<T>(
+    return _showDialog!<T>(
       builder: builder,
       barrierDismissible: barrierDismissible,
       useRootNavigator: useRootNavigator,
@@ -56,7 +55,7 @@ mixin DialogController {
   void dismissSnackBar(
       {SnackBarClosedReason reason = SnackBarClosedReason.hide}) async {
     if (!(await _contextLoaded())) return;
-    Scaffold.of(context).hideCurrentSnackBar(reason: reason);
+    Scaffold.of(context!).hideCurrentSnackBar(reason: reason);
   }
 
   /// Removes the current [SnackBar] by running its normal exit animation.
@@ -65,7 +64,7 @@ mixin DialogController {
   void hideCurrentSnackBar(
       {SnackBarClosedReason reason = SnackBarClosedReason.hide}) async {
     if (!(await _contextLoaded())) return;
-    Scaffold.of(context).hideCurrentSnackBar(reason: reason);
+    Scaffold.of(context!).hideCurrentSnackBar(reason: reason);
   }
 
   /// Removes the current [SnackBar] (if any) immediately.
@@ -75,33 +74,31 @@ mixin DialogController {
   void removeCurrentSnackBar(
       {SnackBarClosedReason reason = SnackBarClosedReason.hide}) async {
     if (!(await _contextLoaded())) return;
-    Scaffold.of(context).removeCurrentSnackBar(reason: reason);
+    Scaffold.of(context!).removeCurrentSnackBar(reason: reason);
   }
 
   /// Shows a [SnackBar] at the bottom of the scaffold.
-  Future<ScaffoldFeatureController<SnackBar, SnackBarClosedReason>>
-      showSnackBar({@required SnackBar Function(BuildContext) builder}) async {
-    assert(builder != null);
+  Future<ScaffoldFeatureController<SnackBar, SnackBarClosedReason>?>
+      showSnackBar({required SnackBar Function(BuildContext?) builder}) async {
     if (!(await _contextLoaded())) return null;
-    return _showSnackBar(builder);
+    return _showSnackBar!(builder);
   }
 
   /// Shows a modal material design bottom sheet.
   ///
   /// A modal bottom sheet is an alternative to a menu or a dialog and prevents
   /// the user from interacting with the rest of the app.
-  Future<T> showModalBottomSheet<T>(
-      {@required Widget Function(BuildContext) builder,
-      Color backgroundColor,
-      double elevation,
-      ShapeBorder shape,
-      Clip clipBehavior,
+  Future<T?> showModalBottomSheet<T>(
+      {required Widget Function(BuildContext) builder,
+      Color? backgroundColor,
+      double? elevation,
+      ShapeBorder? shape,
+      Clip? clipBehavior,
       bool isScrollControlled = false,
       bool useRootNavigator = false,
       bool isDismissible = true}) async {
-    assert(builder != null);
     if (!(await _contextLoaded())) return null;
-    return _showModalBottomSheet<T>(
+    return _showModalBottomSheet!<T>(
         builder: builder,
         backgroundColor: backgroundColor,
         clipBehavior: clipBehavior,
@@ -117,15 +114,14 @@ mixin DialogController {
   ///
   /// Returns a controller that can be used to close and otherwise manipulate the
   /// bottom sheet.
-  Future<PersistentBottomSheetController<T>> showBottomSheet<T>(
-      {@required Widget Function(BuildContext) builder,
-      Color backgroundColor,
-      double elevation,
-      ShapeBorder shape,
-      Clip clipBehavior}) async {
-    assert(builder != null);
+  Future<PersistentBottomSheetController<T>?> showBottomSheet<T>(
+      {required Widget Function(BuildContext) builder,
+      Color? backgroundColor,
+      double? elevation,
+      ShapeBorder? shape,
+      Clip? clipBehavior}) async {
     if (!(await _contextLoaded())) return null;
-    return _showBottomSheet<T>(
+    return _showBottomSheet!<T>(
         builder: builder,
         backgroundColor: backgroundColor,
         elevation: elevation,
@@ -135,32 +131,31 @@ mixin DialogController {
 
   /// Register callbacks
   void registerCallback(
-      {Future<T> Function<T>(
-              {bool barrierDismissible,
-              Widget Function(BuildContext) builder,
-              bool useRootNavigator})
+      {Future<T?> Function<T>(
+              {bool? barrierDismissible,
+              required Widget Function(BuildContext) builder,
+              bool useRootNavigator})?
           showDialog,
-      Future<T> Function<T>(
-              {Widget Function(BuildContext) builder,
-              Color backgroundColor,
-              double elevation,
-              ShapeBorder shape,
-              Clip clipBehavior,
-              bool isScrollControlled,
-              bool useRootNavigator,
-              bool isDismissible})
+      Future<T?> Function<T>(
+              {required Widget Function(BuildContext) builder,
+              Color? backgroundColor,
+              double? elevation,
+              ShapeBorder? shape,
+              Clip? clipBehavior,
+              bool? isScrollControlled,
+              bool? useRootNavigator,
+              bool? isDismissible})?
           showModalBottomSheet,
       ScaffoldFeatureController<SnackBar, SnackBarClosedReason> Function(
-              SnackBar Function(BuildContext) builder)
+              SnackBar Function(BuildContext?) builder)?
           showSnackBar,
       PersistentBottomSheetController<T> Function<T>(
-              {Widget Function(BuildContext) builder,
-              Color backgroundColor,
-              double elevation,
-              ShapeBorder shape,
-              Clip clipBehavior})
+              {Widget Function(BuildContext)? builder,
+              Color? backgroundColor,
+              double? elevation,
+              ShapeBorder? shape,
+              Clip? clipBehavior})?
           showBottomSheet}) {
-    _showDialog = showDialog;
     _showDialog = showDialog;
     _showSnackBar = showSnackBar;
     _showModalBottomSheet = showModalBottomSheet;
@@ -168,9 +163,9 @@ mixin DialogController {
   }
 
   /// Pop the top-most dialog off the OneContext.dialog.
-  popDialog<T extends Object>([T result]) async {
+  popDialog<T extends Object>([T? result]) async {
     if (!(await _contextLoaded())) return;
-    return Navigator.of(context).pop<T>(result);
+    return Navigator.of(context!).pop<T>(result);
   }
 
   Future<bool> _contextLoaded() async {
