@@ -7,56 +7,50 @@ import 'package:flutter/material.dart';
 mixin OverlayController {
   /// Return overlay utility class `OverlayController`
   OverlayController get overlay => this;
-  BuildContext get context => OneContext().context;
+  BuildContext? get context => OneContext().context;
 
-  OverlayEntry _overlayEntry;
+  OverlayEntry? _overlayEntry;
   Map<String, OverlayEntry> _overlays = Map<String, OverlayEntry>();
 
   /// Show the overlay widget keeped by this class (it's like a single instance)
   /// It can be used many times, without an `overlayId`
-  // ignore: missing_return
-  OverlayEntry showOverlay({Widget Function(BuildContext context) builder}) {
+  Future<OverlayEntry> showOverlay(
+      {Widget Function(BuildContext context)? builder}) async {
     hideOverlay();
-    Future.delayed(Duration.zero, () {
-      OverlayState overlayState = Overlay.of(OneContext().context);
-      _overlayEntry = OverlayEntry(builder: builder);
-      overlayState.insert(_overlayEntry);
-      return _overlayEntry;
-    });
+    await Future.delayed(Duration.zero, () {});
+    OverlayState overlayState = Overlay.of(OneContext().context!)!;
+    _overlayEntry = OverlayEntry(builder: builder!);
+    overlayState.insert(_overlayEntry!);
+    return _overlayEntry!;
   }
 
   /// Hide the overlay widget keeped by this class (it's like a single instance)
   /// It can be used many times, without an `overlayId`
   hideOverlay() {
     if (_overlayEntry != null) {
-      _overlayEntry.remove();
+      _overlayEntry!.remove();
       _overlayEntry = null;
     }
   }
 
   /// Add an widget to overlay stack
-  // ignore: missing_return
-  String addOverlay(
-      {@required String overlayId,
-      @required Widget Function(BuildContext context) builder,
-      OverlayEntry below,
-      OverlayEntry above}) {
-    assert(builder != null);
-    assert(overlayId != null);
-
-    Future.delayed(Duration.zero, () {
-      OverlayState overlayState = Overlay.of(OneContext().context);
-      OverlayEntry overlayEntry = OverlayEntry(builder: builder);
-      overlayState.insert(overlayEntry, above: above, below: below);
-      _overlays.putIfAbsent(overlayId, () => overlayEntry);
-      return overlayId;
-    });
+  Future<String> addOverlay(
+      {required String overlayId,
+      required Widget Function(BuildContext context) builder,
+      OverlayEntry? below,
+      OverlayEntry? above}) async {
+    await Future.delayed(Duration.zero, () {});
+    OverlayState overlayState = Overlay.of(OneContext().context!)!;
+    OverlayEntry overlayEntry = OverlayEntry(builder: builder);
+    overlayState.insert(overlayEntry, above: above, below: below);
+    _overlays.putIfAbsent(overlayId, () => overlayEntry);
+    return overlayId;
   }
 
   /// Remove a widget from overlay stack by widget id `overlayId`
   removeOverlay(String overlayId) {
     if (_overlays.containsKey(overlayId)) {
-      OverlayEntry item = _overlays[overlayId];
+      OverlayEntry? item = _overlays[overlayId];
       _overlays.remove(overlayId);
       item?.remove();
       item = null;
@@ -71,7 +65,7 @@ mixin OverlayController {
 
   /// Get OverlayEntry by id, if overlay with this id exists,
   /// else it will return null
-  OverlayEntry getOverlayById(String id) {
+  OverlayEntry? getOverlayById(String id) {
     if (_overlays.containsKey(id)) {
       return _overlays[id];
     }
@@ -96,8 +90,8 @@ mixin OverlayController {
   ///
   /// It is an error to specify both `above` and `below`.
   void rearrange(Iterable<OverlayEntry> newEntries,
-      {OverlayEntry below, OverlayEntry above}) {
-    OverlayState overlayState = Overlay.of(OneContext().context);
+      {OverlayEntry? below, OverlayEntry? above}) {
+    OverlayState overlayState = Overlay.of(OneContext().context!)!;
     overlayState.rearrange(newEntries, below: below, above: above);
   }
 
@@ -108,17 +102,17 @@ mixin OverlayController {
   /// asserts. To avoid people depending on it, this function is implemented
   /// only in debug mode, and always returns false in release mode.
   bool debugIsVisible(OverlayEntry entry) {
-    OverlayState overlayState = Overlay.of(OneContext().context);
+    OverlayState overlayState = Overlay.of(OneContext().context!)!;
     return overlayState.debugIsVisible(entry);
   }
 
   /// Show circular progress indicator, or a custom widget.
   /// It can be used generally to show progress indicator,
   /// but you can use a completelly custom widget
-  OverlayEntry showProgressIndicator(
-      {Widget Function(BuildContext context) builder,
-      Color backgroundColor,
-      Color circularProgressIndicatorColor}) {
+  Future<OverlayEntry> showProgressIndicator(
+      {Widget Function(BuildContext? context)? builder,
+      Color? backgroundColor,
+      Color? circularProgressIndicatorColor}) {
     return showOverlay(
         builder: (_) => Stack(
               children: <Widget>[
